refactor: rename generator-core to generateProject and split into seams
The 140-line generateProject function tangled four distinct concerns:
file walking, placeholder substitution, file writing, and shell-out
post-install steps. The trickiest logic — the -keep file/dir rename
rules combined with placeholder detection — was reachable only by
running the full generator against a real filesystem.

This change preserves the external interface (`generateProject(options)`
still does end-to-end generation) but extracts three internal seams:

  1. SourceReader — a port over "list files in a dir" + "read a file as
     text". Production adapter wraps Bun.Glob + Bun.file. The test-only
     adapter (added in the next commit) wraps a Map<path, contents>, so
     planTemplate can be exercised in-memory.

  2. planTemplate(args) — pure given a SourceReader and a frozen
     replacements map. Returns FileWrite[] = (destPath, contents) tuples
     with all -keep renames and placeholder substitutions already
     applied. No disk writes, no shell, no exits.

     Source dirs are walked in order; later sources win on
     destination-path conflict (Map<destPath, FileWrite> by construction).
     This formalises the previous loop's behaviour, where the second
     source dir would overwrite the first by virtue of running second.

  3. applyPlan(plan) — writes the planned tuples to disk, creating
     parent dirs as needed. Trivial; lives separately so plan tests
     don't pay for I/O.

  4. runPostInstall(targetDir) — the side-effecting tail: bun install,
     biomeInit, optional .vscode/settings.json merge, git init. Stays
     imperative; git failure stays non-fatal (warning, not throw).

generateProject orchestrates: build replacements (including the random
BETTER_AUTH_SECRET), call planTemplate, applyPlan, runPostInstall.

The file is renamed to generateProject.ts to match the export. The old
"-core" suffix was meaningful when cli-mode and guided-mode were the
"non-core" parts; once those collapse into resolveProjectOptions in a
later step, the suffix loses its referent.

planTemplate, applyPlan, SourceReader, and FileWrite are exported so
the next commit's tests can import them directly. Per the deepening
language: these are *internal seams* exposed for the module's own tests
— callers outside this module should still consume only generateProject.

Verified:
  - bun test: 26 pass / 4 fail (same pre-existing failures)
  - bun run check: 3 errors / 2 warnings (same pre-existing, all in
    template files)
  - smoke gen with --type fullstack --yes: package.json name correctly
    substituted, file tree intact
